Initial refactoring and removal of examples
